// libs/integration-tests/src/lib/models/container.enum.ts
export enum CONTAINER {
  POSTGRES = 'postgresdb',
  KEYCLOAK = 'keycloak-app',
  IAMKC_SVC = 'onecx-iam-kc-svc',
  WORKSPACE_SVC = 'onecx-workspace-svc',
  USER_PROFILE_SVC = 'onecx-user-profile-svc',
  THEME_SVC = 'onecx-theme-svc',
  TENANT_SVC = 'onecx-tenant-svc',
  PRODUCT_STORE_SVC = 'onecx-product-store-svc',
  PERMISSION_SVC = 'onecx-permission-svc',
  SHELL_BFF = 'onecx-shell-bff',
  SHELL_UI = 'onecx-shell-ui',
  IMPORT_MANAGER = 'importManager',
}